Revert "add async file write support to files archive"
This reverts commit b8cfa56977d98520aa8b62252c3a2726d440afe0.
