+++ planetiler-core/src/main/java/com/onthegomap/planetiler/stats/ProcessInfo.java
@@ -234,7 +234,6 @@ public class ProcessInfo {
   public static ThreadState getCurrentThreadState() {
     ThreadMXBean threadMXBean = ManagementFactory.getThreadMXBean();
     ThreadInfo thread = threadMXBean.getThreadInfo(Thread.currentThread().getId());
-    // null for virtual - tmp workaround
-    return thread == null ? null : new ThreadState(threadMXBean, thread);
+    return new ThreadState(threadMXBean, thread);
   }
 }
