+++ planetiler-core/src/main/java/com/onthegomap/planetiler/stats/Timers.java
@@ -115,11 +115,7 @@ public class Timers {
   public void finishedWorker(String prefix, Duration elapsed) {
     Stage stage = currentStage.get();
     if (stage != null) {
-      final ProcessInfo.ThreadState threadState = ProcessInfo.getCurrentThreadState();
-      // null for virtual - tmp workaround
-      if (threadState != null) {
-        stage.threadStats.add(new ThreadInfo(threadState, prefix, elapsed));
-      }
+      stage.threadStats.add(new ThreadInfo(ProcessInfo.getCurrentThreadState(), prefix, elapsed));
     }
   }
