--- planetiler-core/src/main/java/com/onthegomap/planetiler/worker/ExecutorType.java
@@ -1,6 +1,0 @@
-package com.onthegomap.planetiler.worker;
-
-public enum ExecutorType {
-  FIXED,
-  VIRTUAL,
-}
